little refactor / ResultAndSort renamed to TopBar
